GlShader: Fix configuration of multiple sampler uniforms
OpenGL sampler uniforms need to be set to the index of the texture unit from
which they should be reading. We by default allocate texture units in order, one
for each sampler.

We however didn't actually set the uniform value, so while everything was
working fine with a single sampler (uniforms are initialized to 0 by default),
the moment a shader used multiple samplers, they would all try to read from the
first texture unit instead of their assigned unit.

This commit fixes this by properly initializing the uniform value to the correct
texture unit when the sampler uniform object is created.